requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyYAML==6.0.2
SQLAlchemy==2.0.35
pymysql==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
email-validator==2.1.1
